requirements.txt
streamlit>=1.37.0
streamlit-aggrid>=0.3.4
streamlit-chatbox
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
pyyaml>=6.0.1
python-dotenv>=1.0.1
requests>=2.31.0
plotly>=5.22.0

gspread>=5.12.0
oauth2client>=4.1.3

langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

openai>=1.30.0
tiktoken>=0.7.0

pinecone-client>=3.2.0
langchain-pinecone>=0.1.0
faiss-cpu>=1.8.0
sentence-transformers>=2.2.2
psycopg2_binary==2.9.11
SQLAlchemy==2.0.44
pydantic==2.12.2

networkx>=3.3
beautifulsoup4>=4.12.3
nltk>=3.8.1
unstructured>=0.14.4
tabulate